fix(fb): z_mixture_kernel took an `Int` — nightly rejects it as a kernel arg
`fb_train_gpu.mojo` stopped compiling on NVIDIA:

    constraint failed: Int and UInt do not conform to DevicePassable;
    use a fixed-width type such as Int32 or Int64 instead

`z_mixture_kernel`'s `n_b_rows` was `Int` and had been since it was
written; the toolchain tightened underneath it. A runtime scalar argument
to a GPU kernel must now be fixed-width.

⚠ The diagnostic points at the CALL SITE, not the kernel. The failure
surfaces as a chain of "function instantiation failed" notes descending
through `enqueue_function` -> device_context -> simd.mojo, and only the
last line names the real constraint — nothing in it mentions
`n_b_rows` or its declaration. Audited every kernel in `deep_agents/fb/`
and `nn/primitives/pairwise_dot.mojo` by signature: this was the only
one.

The parameter is `Int32` at the ABI and widened to `Int` once inside the
body, so the comparisons against the row index stay Int-typed. Gated by
rebuilding `fb_train_gpu.mojo` (the constraint is target-independent, so
an Apple build exercises the same instantiation).

+++ examples/fb/fb_train_gpu.mojo
@@ -288,7 +288,7 @@ def main() raises:
             t.b_sp.dev.value().unsafe_ptr(),
             pick.dev.value().unsafe_ptr(),
             Scalar[DT](0.5),
-            BATCH,
+            Int32(BATCH),  # fixed-width: `Int` is not DevicePassable
             grid_dim=_blocks(BATCH),
             block_dim=TPB,
         )
